requirements.txt
flask>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.5.0
pillow>=8.0.0
easyocr>=1.6.0
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
